PYIC-1718: Rename credential-issuer-config lambda
