+++ lambdas/get-credential-issuer-config/build.gradle
