+++ lambdas/get-credential-issuer-config/src/main/resources/log4j2.yaml
