+++ settings.gradle
@@ -4,7 +4,7 @@ include "lib",
 		"lambdas:issue-client-access-token",
 		"lambdas:sessionend",
 		"lambdas:credentialissuerreturn",
-		"lambdas:credentialissuerconfig",
+		"lambdas:get-credential-issuer-config",
 		"lambdas:evaluate-gpg45-scores",
 		"lambdas:issuedcredentials",
 		"lambdas:sessionstart",
